Moving nav js to new helper file
